Make `ouroboros-network-api` independent of `ouroboros-network-framework`
This way `ouroboros-network-protocols` does not contain
`ouroboros-network-framework` as its transitive dependency.
